# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Hospital AI - 管理端","version":"0.1.0"},"paths":{"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Health Get"}}}}}}},"/operation-journals/":{"get":{"tags":["operation_journals"],"summary":"Query Operation Journal","operationId":"query_operation_journal_operation_journals__get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","description":"Page size","default":20,"title":"Page Size"},"description":"Page size"},{"name":"id_eq","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id Eq"}},{"name":"id_in","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id In"}},{"name":"trace_id_eq","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Trace Id Eq"}},{"name":"trace_id_in","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Trace Id In"}},{"name":"user_id_eq","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id Eq"}},{"name":"user_id_in","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id In"}},{"name":"user_role_eq","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Role Eq"}},{"name":"user_role_in","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Role In"}},{"name":"action_code_eq","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Action Code Eq"}},{"name":"action_code_in","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Action Code In"}},{"name":"time_eq","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Time Eq"}},{"name":"time_lt","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Time Lt"}},{"name":"time_gt","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Time Gt"}},{"name":"time_in","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Time In"}},{"name":"time_gte","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Time Gte"}},{"name":"time_lte","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Time Lte"}},{"name":"status_eq","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status Eq"}},{"name":"status_in","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status In"}},{"name":"target_id_eq","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Id Eq"}},{"name":"target_id_in","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Id In"}},{"name":"target_role_eq","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Role Eq"}},{"name":"target_role_in","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Role In"}},{"name":"sort_by","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Allowed values: id, time, descend with prefix '-'","title":"Sort By"},"description":"Allowed values: id, time, descend with prefix '-'"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResult_OperationJournal_"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/operation-journals/{id}":{"delete":{"tags":["operation_journals"],"summary":"Delete Operation Journal","operationId":"delete_operation_journal_operation_journals__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OperationJournal":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"trace_id":{"type":"string","format":"uuid","title":"Trace Id"},"source":{"type":"string","title":"Source"},"user_id":{"type":"string","title":"User Id"},"user_role":{"type":"string","title":"User Role"},"action_code":{"type":"string","title":"Action Code"},"time":{"type":"string","format":"date-time","title":"Time"},"status":{"type":"integer","title":"Status"},"is_deleted":{"type":"boolean","title":"Is Deleted","default":false},"target_id":{"type":"string","title":"Target Id"},"target_role":{"type":"string","title":"Target Role"},"payload":{"additionalProperties":true,"type":"object","title":"Payload"}},"type":"object","required":["trace_id","source","user_id","user_role","action_code","status","target_id","target_role"],"title":"OperationJournal"},"PaginatedResult_OperationJournal_":{"properties":{"total_count":{"type":"integer","title":"Total Count"},"page_count":{"type":"integer","title":"Page Count"},"page":{"type":"integer","title":"Page"},"page_size":{"type":"integer","title":"Page Size"},"data":{"items":{"$ref":"#/components/schemas/OperationJournal"},"type":"array","title":"Data"}},"type":"object","required":["total_count","page_count","page","page_size","data"],"title":"PaginatedResult[OperationJournal]"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}